+++ services/storage/tests/fixtures/data_models.py
@@ -3,19 +3,27 @@
 # pylint:disable=redefined-outer-name
 
 
+from collections import deque
 from contextlib import asynccontextmanager
+from random import choice, randint
 from typing import Any, AsyncIterator, Awaitable, Callable
 
 import pytest
 import sqlalchemy as sa
 from aiopg.sa.connection import SAConnection
 from aiopg.sa.engine import Engine
+from anyio import Path
 from faker import Faker
+from models_library.basic_types import SHA256Str
 from models_library.projects import ProjectID
-from models_library.projects_nodes_io import NodeID
+from models_library.projects_nodes_io import NodeID, SimcoreS3FileID
 from models_library.users import UserID
+from pydantic import ByteSize, parse_obj_as
 from pytest_simcore.helpers.rawdata_fakers import random_project, random_user
+from servicelib.utils import logged_gather
 from simcore_postgres_database.storage_models import projects, users
+
+from ..helpers.utils import get_updated_project
 
 
 @asynccontextmanager
@@ -167,3 +175,117 @@ async def create_project_node(
         return new_node_id
 
     yield _creator
+
+
+@pytest.fixture
+async def random_project_with_files(
+    aiopg_engine: Engine,
+    create_project: Callable[[], Awaitable[dict[str, Any]]],
+    create_project_node: Callable[..., Awaitable[NodeID]],
+    create_simcore_file_id: Callable[
+        [ProjectID, NodeID, str, Path | None], SimcoreS3FileID
+    ],
+    upload_file: Callable[..., Awaitable[tuple[Path, SimcoreS3FileID]]],
+    faker: Faker,
+) -> Callable[
+    [int, tuple[ByteSize, ...]],
+    Awaitable[
+        tuple[
+            dict[str, Any], dict[NodeID, dict[SimcoreS3FileID, dict[str, Path | str]]]
+        ]
+    ],
+]:
+    async def _creator(
+        num_nodes: int = 12,
+        file_sizes: tuple[ByteSize, ...] = (
+            parse_obj_as(ByteSize, "7Mib"),
+            parse_obj_as(ByteSize, "110Mib"),
+            parse_obj_as(ByteSize, "1Mib"),
+        ),
+        file_checksums: tuple[SHA256Str, ...] = (
+            parse_obj_as(
+                SHA256Str,
+                "311e2e130d83cfea9c3b7560699c221b0b7f9e5d58b02870bd52b695d8b4aabd",
+            ),
+            parse_obj_as(
+                SHA256Str,
+                "08e297db979d3c84f6b072c2a1e269e8aa04e82714ca7b295933a0c9c0f62b2e",
+            ),
+            parse_obj_as(
+                SHA256Str,
+                "488f3b57932803bbf644593bd46d95599b1d4da1d63bc020d7ebe6f1c255f7f3",
+            ),
+        ),
+    ) -> tuple[
+        dict[str, Any], dict[NodeID, dict[SimcoreS3FileID, dict[str, Path | str]]]
+    ]:
+        assert len(file_sizes) == len(file_checksums)
+        project = await create_project()
+        src_projects_list: dict[
+            NodeID, dict[SimcoreS3FileID, dict[str, Path | str]]
+        ] = {}
+        upload_tasks: deque[Awaitable] = deque()
+        for _node_index in range(num_nodes):
+            # NOTE: we put some more outputs in there to simulate a real case better
+            new_node_id = NodeID(faker.uuid4())
+            output3_file_id = create_simcore_file_id(
+                ProjectID(project["uuid"]),
+                new_node_id,
+                faker.file_name(),
+                Path("outputs/output3"),
+            )
+            src_node_id = await create_project_node(
+                ProjectID(project["uuid"]),
+                new_node_id,
+                outputs={
+                    "output_1": faker.pyint(),
+                    "output_2": faker.pystr(),
+                    "output_3": f"{output3_file_id}",
+                },
+            )
+            assert src_node_id == new_node_id
+
+            # upload the output 3 and some random other files at the root of each node
+            src_projects_list[src_node_id] = {}
+            checksum: SHA256Str = choice(file_checksums)
+            src_file, _ = await upload_file(
+                file_size=choice(file_sizes),
+                file_name=Path(output3_file_id).name,
+                file_id=output3_file_id,
+                sha256_checksum=checksum,
+            )
+            src_projects_list[src_node_id][output3_file_id] = {
+                "path": src_file,
+                "sha256_checksum": checksum,
+            }
+
+            async def _upload_file_and_update_project(project, src_node_id):
+                src_file_name = faker.file_name()
+                src_file_uuid = create_simcore_file_id(
+                    ProjectID(project["uuid"]), src_node_id, src_file_name, None
+                )
+                checksum: SHA256Str = choice(file_checksums)
+                src_file, _ = await upload_file(
+                    file_size=choice(file_sizes),
+                    file_name=src_file_name,
+                    file_id=src_file_uuid,
+                    sha256_checksum=checksum,
+                )
+                src_projects_list[src_node_id][src_file_uuid] = {
+                    "path": src_file,
+                    "sha256_checksum": checksum,
+                }
+
+            # add a few random files in the node storage
+            upload_tasks.extend(
+                [
+                    _upload_file_and_update_project(project, src_node_id)
+                    for _ in range(randint(0, 3))
+                ]
+            )
+        await logged_gather(*upload_tasks, max_concurrency=2)
+
+        project = await get_updated_project(aiopg_engine, project["uuid"])
+        return project, src_projects_list
+
+    return _creator
