+++ services/storage/tests/unit/conftest.py
@@ -8,8 +8,6 @@ from collections import deque
 from collections.abc import AsyncIterator, Awaitable, Callable
 from contextlib import AbstractAsyncContextManager, asynccontextmanager
 from pathlib import Path
-from urllib.parse import quote
-from uuid import UUID
 
 import openapi_core
 import pytest
@@ -218,20 +216,20 @@ async def with_versioning_enabled(
     )
 
 
-@pytest.fixture
-async def upload_file(client: TestClient, user_id: UserID, location_id: LocationID):
-    async def _(file: Path, file_id: UUID):
-        assert file.is_file()
-        assert client.app
-        url = (
-            client.app.router["upload_file"]
-            .url_for(
-                location_id=f"{location_id}",
-                file_id=quote(f"api/{file_id}/{file.name}", safe=""),
-            )
-            .with_query(file_size=f"{file.stat().st_size}", user_id=f"{user_id}")
-        )
-        response = await client.put(f"{url}")
-        assert response.status == 200
+# @pytest.fixture
+# async def upload_file(client: TestClient, user_id: UserID, location_id: LocationID):
+#     async def _(file: Path, file_id: UUID):
+#         assert file.is_file()
+#         assert client.app
+#         url = (
+#             client.app.router["upload_file"]
+#             .url_for(
+#                 location_id=f"{location_id}",
+#                 file_id=quote(f"api/{file_id}/{file.name}", safe=""),
+#             )
+#             .with_query(file_size=f"{file.stat().st_size}", user_id=f"{user_id}")
+#         )
+#         response = await client.put(f"{url}")
+#         assert response.status == 200
 
-    return _
+#     return _
