+++ services/storage/tests/unit/test_handlers_simcore_s3.py
@@ -6,11 +6,8 @@
 # pylint:disable=too-many-nested-blocks
 
 import sys
-from collections import deque
 from copy import deepcopy
 from pathlib import Path
-from random import randint
-from secrets import choice
 from typing import Any, Awaitable, Callable, Literal
 
 import pytest
@@ -39,6 +36,8 @@ from simcore_service_storage.simcore_s3_dsm import SimcoreS3DataManager
 from tests.helpers.utils_file_meta_data import assert_file_meta_data_in_db
 from tests.helpers.utils_project import clone_project_data
 from yarl import URL
+
+from ..helpers.utils import get_updated_project
 
 pytest_simcore_core_services_selection = ["postgres"]
 pytest_simcore_ops_services_selection = ["adminer"]
@@ -178,130 +177,6 @@ async def test_copy_folders_from_empty_project(
         assert num_entries == 0
 
 
-async def _get_updated_project(aiopg_engine: Engine, project_id: str) -> dict[str, Any]:
-    async with aiopg_engine.acquire() as conn:
-        result = await conn.execute(
-            sa.select(projects).where(projects.c.uuid == project_id)
-        )
-        row = await result.fetchone()
-        assert row
-        return dict(row)
-
-
-@pytest.fixture
-async def random_project_with_files(
-    aiopg_engine: Engine,
-    create_project: Callable[[], Awaitable[dict[str, Any]]],
-    create_project_node: Callable[..., Awaitable[NodeID]],
-    create_simcore_file_id: Callable[
-        [ProjectID, NodeID, str, Path | None], SimcoreS3FileID
-    ],
-    upload_file: Callable[..., Awaitable[tuple[Path, SimcoreS3FileID]]],
-    faker: Faker,
-) -> Callable[
-    [int, tuple[ByteSize, ...]],
-    Awaitable[
-        tuple[
-            dict[str, Any], dict[NodeID, dict[SimcoreS3FileID, dict[str, Path | str]]]
-        ]
-    ],
-]:
-    async def _creator(
-        num_nodes: int = 12,
-        file_sizes: tuple[ByteSize, ...] = (
-            parse_obj_as(ByteSize, "7Mib"),
-            parse_obj_as(ByteSize, "110Mib"),
-            parse_obj_as(ByteSize, "1Mib"),
-        ),
-        file_checksums: tuple[SHA256Str, ...] = (
-            parse_obj_as(
-                SHA256Str,
-                "311e2e130d83cfea9c3b7560699c221b0b7f9e5d58b02870bd52b695d8b4aabd",
-            ),
-            parse_obj_as(
-                SHA256Str,
-                "08e297db979d3c84f6b072c2a1e269e8aa04e82714ca7b295933a0c9c0f62b2e",
-            ),
-            parse_obj_as(
-                SHA256Str,
-                "488f3b57932803bbf644593bd46d95599b1d4da1d63bc020d7ebe6f1c255f7f3",
-            ),
-        ),
-    ) -> tuple[
-        dict[str, Any], dict[NodeID, dict[SimcoreS3FileID, dict[str, Path | str]]]
-    ]:
-        assert len(file_sizes) == len(file_checksums)
-        project = await create_project()
-        src_projects_list: dict[
-            NodeID, dict[SimcoreS3FileID, dict[str, Path | str]]
-        ] = {}
-        upload_tasks: deque[Awaitable] = deque()
-        for _node_index in range(num_nodes):
-            # NOTE: we put some more outputs in there to simulate a real case better
-            new_node_id = NodeID(faker.uuid4())
-            output3_file_id = create_simcore_file_id(
-                ProjectID(project["uuid"]),
-                new_node_id,
-                faker.file_name(),
-                Path("outputs/output3"),
-            )
-            src_node_id = await create_project_node(
-                ProjectID(project["uuid"]),
-                new_node_id,
-                outputs={
-                    "output_1": faker.pyint(),
-                    "output_2": faker.pystr(),
-                    "output_3": f"{output3_file_id}",
-                },
-            )
-            assert src_node_id == new_node_id
-
-            # upload the output 3 and some random other files at the root of each node
-            src_projects_list[src_node_id] = {}
-            checksum: SHA256Str = choice(file_checksums)
-            src_file, _ = await upload_file(
-                file_size=choice(file_sizes),
-                file_name=Path(output3_file_id).name,
-                file_id=output3_file_id,
-                sha256_checksum=checksum,
-            )
-            src_projects_list[src_node_id][output3_file_id] = {
-                "path": src_file,
-                "sha256_checksum": checksum,
-            }
-
-            async def _upload_file_and_update_project(project, src_node_id):
-                src_file_name = faker.file_name()
-                src_file_uuid = create_simcore_file_id(
-                    ProjectID(project["uuid"]), src_node_id, src_file_name, None
-                )
-                checksum: SHA256Str = choice(file_checksums)
-                src_file, _ = await upload_file(
-                    file_size=choice(file_sizes),
-                    file_name=src_file_name,
-                    file_id=src_file_uuid,
-                    sha256_checksum=checksum,
-                )
-                src_projects_list[src_node_id][src_file_uuid] = {
-                    "path": src_file,
-                    "sha256_checksum": checksum,
-                }
-
-            # add a few random files in the node storage
-            upload_tasks.extend(
-                [
-                    _upload_file_and_update_project(project, src_node_id)
-                    for _ in range(randint(0, 3))
-                ]
-            )
-        await logged_gather(*upload_tasks, max_concurrency=2)
-
-        project = await _get_updated_project(aiopg_engine, project["uuid"])
-        return project, src_projects_list
-
-    return _creator
-
-
 @pytest.fixture
 def short_dsm_cleaner_interval(monkeypatch: pytest.MonkeyPatch) -> int:
     monkeypatch.setenv("STORAGE_CLEANER_INTERVAL_S", "1")
@@ -346,7 +221,7 @@ async def test_copy_folders_from_valid_project_with_one_large_file(
         nodes_map={NodeID(i): NodeID(j) for i, j in nodes_map.items()},
     )
     assert data == jsonable_encoder(
-        await _get_updated_project(aiopg_engine, dst_project["uuid"])
+        await get_updated_project(aiopg_engine, dst_project["uuid"])
     )
     # check that file meta data was effectively copied
     for src_node_id in src_projects_list:
@@ -404,7 +279,7 @@ async def test_copy_folders_from_valid_project(
         nodes_map={NodeID(i): NodeID(j) for i, j in nodes_map.items()},
     )
     assert data == jsonable_encoder(
-        await _get_updated_project(aiopg_engine, dst_project["uuid"])
+        await get_updated_project(aiopg_engine, dst_project["uuid"])
     )
 
     # check that file meta data was effectively copied
